refactor: extract SessionStorage into separate module
Separates session storage logic from SessionManager to improve code
organization and maintainability. This is phase 1 of the session
management refactoring.

Changes:
- Extract SessionStorage into new session_storage.rs module
- Keep SessionManager as thin wrapper using singleton pattern
- Move constants (CURRENT_SCHEMA_VERSION, SESSIONS_FOLDER, DB_NAME) to session_storage
- Update imports in posthog.rs and goose-cli info.rs
- Retain search_chat_history in SessionManager (used by chatrecall)
- Remove import_legacy_sessions method as planned

All existing tests pass. Test coverage for new functions will be added
in follow-up PRs.
