+++ crates/goose/src/session/mod.rs
@@ -3,7 +3,9 @@ mod diagnostics;
 pub mod extension_data;
 mod legacy;
 pub mod session_manager;
+pub mod session_storage;
 
 pub use diagnostics::{generate_diagnostics, get_system_info, SystemInfo};
 pub use extension_data::{EnabledExtensionsState, ExtensionData, ExtensionState, TodoState};
 pub use session_manager::{Session, SessionInsights, SessionManager, SessionType};
+pub use session_storage::SessionStorage;
